requirements.txt
Flask==2.3.3
pycryptodome==3.19.0
gunicorn==21.2.0 